Redacted undesired changes of b53f2d6271c
Problems with the editor lead to unintended input of characters
creating a critical bug in one place that prevented the propper
read of the configuration if an device offset had to be read. It
unfortunately did not attract attention in the patch and former
tests.

This fix has been tested with RTU/TCP master and TCP slave on
buster.