Establish test pipeline for static code analysis
Uses docker that is pulled from container registry. Due to issue
https://github.com/actions/runner/issues/434 in github actions
there is no clean sleeve at the start of the run, actions do not
use temporary workspaces but employ existing file systems on the
runner's system instead. This can lead to undefined behavior if
no cleanup is executed before checkout. Also mind that there
will be issues with the file system's permission if the docker
container's user does not match the UID of the action runner's UID.