fix: Daemons does not start after upgrade
In Commit d054221, we changed from dh_systemd_enable to
dh_installsystemd. The parameter —-no-start was added. This should
prevent a start on the first installation of the package, but it also
prevent a restart after package upgrade. Our unit file is installed
with -—no-enabled. This sets the unit file to disabled during
installation and dh-systemd-invoke will NOT start the program at the
first installation or upgrade until the user sets it to enabled. We can
and must remove —-no-start.

Signed-off-by: Sven Sager <s.sager@kunbus.com>